2020-Bot-Discord/rodrigoaramburu/requirements.txt
discord.py
python-decouple
pymongo
feedparser
google-api-python-client
tabulate
googletrans